Fix for https://issues.apache.org/activemq/browse/AMQ-1167. Thanks for the heads up Woonsan. Added some spelling corrections.
git-svn-id: https://svn.apache.org/repos/asf/activemq/branches/activemq-4.1@548207 13f79535-47bb-0310-9956-ffa450edef68
